feature/Added Password reset using OTP.

+++ security/src/main/java/com/fincity/security/dto/policy/ClientPinPolicy.java
@@ -2,9 +2,6 @@ package com.fincity.security.dto.policy;
 
 import java.io.Serial;
 
-import org.jooq.types.ULong;
-import org.jooq.types.UShort;
-
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fincity.saas.commons.util.CodeUtil;
@@ -23,19 +20,19 @@ public class ClientPinPolicy extends AbstractPolicy {
 	@Serial
 	private static final long serialVersionUID = 6320470382858314209L;
 
-	private UShort length = UShort.valueOf(4);
+	private Short length = 4;
 
 	@JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
-	private ULong reLoginAfterInterval = ULong.valueOf(120);
+	private Long reLoginAfterInterval = 120L;
 
 	@JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
-	private UShort expiryInDays = UShort.valueOf(30);
+	private Short expiryInDays = 30;
 
 	@JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
-	private UShort expiryWarnInDays = UShort.valueOf(25);
+	private Short expiryWarnInDays = 25;
 
 	@JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
-	private UShort pinHistoryCount = UShort.valueOf(3);
+	private Short pinHistoryCount = 3;
 
 	@Override
 	@JsonIgnore
